add default implementation of IDevicePortalConnection
