+++ WindowsDevicePortalWrapper/WindowsDevicePortalWrapper.UniversalWindows/WindowsDevicePortalWrapper.UniversalWindows.csproj
@@ -110,6 +110,7 @@
   <ItemGroup>
     <Compile Include="CertificateHandling.cs" />
     <Compile Include="Core\AppDeployment.cs" />
+    <Compile Include="DefaultDevicePortalConnection.cs" />
     <Compile Include="HttpRest\RestPost.cs" />
     <Compile Include="HttpRest\RestPut.cs" />
     <Compile Include="HttpRest\RestDelete.cs" />
