+++ WindowsDevicePortalWrapper/WindowsDevicePortalWrapper/DefaultDevicePortalConnection.cs
@@ -1,5 +1,5 @@
 ﻿//----------------------------------------------------------------------------------------------
-// <copyright file="DevicePortalConnection.cs" company="Microsoft Corporation">
+// <copyright file="DefaultDevicePortalConnection.cs" company="Microsoft Corporation">
 //     Licensed under the MIT License. See LICENSE.TXT in the project root license information.
 // </copyright>
 //----------------------------------------------------------------------------------------------
@@ -8,38 +8,38 @@ using System;
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using System.Text.RegularExpressions;
-using Microsoft.Tools.WindowsDevicePortal;
 using static Microsoft.Tools.WindowsDevicePortal.DevicePortal;
 
-namespace MockDataGenerator
+namespace Microsoft.Tools.WindowsDevicePortal
 {
     /// <summary>
-    /// IDevicePortalConnection implementation for MockDataGenerator test project
+    /// Default implementation of the IDevicePortalConnection interface.
+    /// This implementation is designed to be compatibile with all device families.
     /// </summary>
-    public class DevicePortalConnection : IDevicePortalConnection
+    public class DefaultDevicePortalConnection : IDevicePortalConnection
     {
         /// <summary>
-        /// Device Certificate
+        /// The device's root certificate.
         /// </summary>
         private X509Certificate2 deviceCertificate = null;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="DevicePortalConnection"/> class.
+        /// Initializes a new instance of the <see cref="DevicePortalConnection" /> class.
         /// </summary>
-        /// <param name="address">device identifier</param>
-        /// <param name="userName">WDP username</param>
-        /// <param name="password">WDP password</param>
-        public DevicePortalConnection(
+        /// <param name="address">The fully qualified (ex: "https:/1.2.3.4:4321") address of the device.</param>
+        /// <param name="userName">The user name used in the connection credentials.</param>
+        /// <param name="password">The password used in the connection credentials.</param>
+        public DefaultDevicePortalConnection(
             string address,
             string userName,
             string password)
         {
-            this.Connection = new Uri(string.Format("https://{0}", address));
+            this.Connection = new Uri(address);
             this.Credentials = new NetworkCredential(userName, password);
         }
 
         /// <summary>
-        /// Gets Connection property
+        /// Gets the URI used to connect to the device.
         /// </summary>
         public Uri Connection
         {
@@ -61,7 +61,7 @@ namespace MockDataGenerator
 
                 string absoluteUri = this.Connection.AbsoluteUri;
 
-                if (absoluteUri.StartsWith("https", StringComparison.InvariantCultureIgnoreCase))
+                if (absoluteUri.StartsWith("https", StringComparison.OrdinalIgnoreCase))
                 {
                     return new Uri(Regex.Replace(absoluteUri, "https", "wss", RegexOptions.IgnoreCase | RegexOptions.CultureInvariant));
                 }
@@ -73,7 +73,7 @@ namespace MockDataGenerator
         }
 
         /// <summary>
-        /// Gets Credentials property
+        /// Gets the credentials used to connect to the device.
         /// </summary>
         public NetworkCredential Credentials
         {
@@ -82,16 +82,16 @@ namespace MockDataGenerator
         }
 
         /// <summary>
-        /// Gets or sets device family
+        /// Gets or sets the device's operating system family.
         /// </summary>
         public string Family
-        { 
-            get; 
+        {
+            get;
             set;
         }
 
         /// <summary>
-        /// Gets or sets the device name
+        /// Gets or sets the device name.
         /// </summary>
         public string Name
         {
@@ -100,7 +100,7 @@ namespace MockDataGenerator
         }
 
         /// <summary>
-        /// Gets or sets device OS Info
+        /// Gets or sets the operating system information.
         /// </summary>
         public OperatingSystemInformation OsInfo
         {
@@ -109,9 +109,9 @@ namespace MockDataGenerator
         }
 
         /// <summary>
-        /// Returns certificate data
+        /// Gets the raw device certificate.
         /// </summary>
-        /// <returns>certificate data</returns>
+        /// <returns>Byte array containing the raw certificate data.</returns>
         public byte[] GetDeviceCertificateData()
         {
             return this.deviceCertificate.GetRawCertData();
@@ -136,41 +136,51 @@ namespace MockDataGenerator
         }
 
         /// <summary>
-        /// Sets certificate data
+        /// Updates the device's connection Uri.
         /// </summary>
-        /// <param name="certificateData">certificate data</param>
-        public void SetDeviceCertificate(byte[] certificateData)
-        {
-            X509Certificate2 cert = new X509Certificate2(certificateData);
-            if (!cert.IssuerName.Name.Contains(DevicePortalCertificateIssuer))
-            {
-                throw new DevicePortalException(
-                    (HttpStatusCode)0,
-                    "Invalid certificate issuer",
-                    null,
-                    "Failed to download device certificate");
-            }
-
-            this.deviceCertificate = cert;
-        }
-
-        /// <summary>
-        /// MockDataGenerator will never update the connection.
-        /// </summary>
-        /// <param name="requiresHttps">https required</param>
+        /// <param name="requiresHttps">Indicates whether or not to always require a secure connection.</param>
         public void UpdateConnection(bool requiresHttps)
         {
-            throw new NotImplementedException();
+            this.Connection = new Uri(
+                string.Format(
+                    "{0}://{1}",
+                    requiresHttps ? "https" : "http",
+                    this.Connection.Authority));
         }
-
+        
         /// <summary>
-        ///  MockDataGenerator will never update the connection.
+        /// Updates the device's connection Uri.
         /// </summary>
-        /// <param name="ipConfig">IP info</param>
-        /// <param name="requiresHttps">https required</param>
-        public void UpdateConnection(IpConfiguration ipConfig, bool requiresHttps)
+        /// <param name="ipConfig">The device's IP configuration data.</param>
+        /// <param name="requiresHttps">Indicates whether or not to always require a secure connection.</param>
+        public void UpdateConnection(
+            IpConfiguration ipConfig,
+            bool requiresHttps = false)
         {
-            throw new NotImplementedException();
+            Uri newConnection = null;
+
+            foreach (NetworkAdapterInfo adapter in ipConfig.Adapters)
+            {
+                foreach (IpAddressInfo addressInfo in adapter.IpAddresses)
+                {
+                    // We take the first, non-169.x.x.x address we find that is not 0.0.0.0.
+                    if ((addressInfo.Address != "0.0.0.0") && !addressInfo.Address.StartsWith("169."))
+                    {
+                        newConnection = new Uri(
+                            string.Format(
+                                "{0}://{1}", 
+                                requiresHttps ? "https" : "http",
+                                this.Connection.Authority));
+                        break;
+                    }
+                }
+
+                if (newConnection != null)
+                {
+                    this.Connection = newConnection;
+                    break;
+                }
+            }
         }
     }
 }
